+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## 2.0.21 (29.11.2014)
 * [Add] Добавлен класс \Helpers\Collection для работы с коллекциями
 * [Fix] Баг в \Helpers\FS::getInexistantFilename() (Issue #106)
+* [Fix] Отсутствие ID записи в modResource::toArray() (Issue #104)
 * [Refactor] Доработан алгоритм метода \Helpers\FS::takeFileMIME()
 
 ## 2.0.20 (28.11.2014)